Adding Watch, Colliders, Prefabs and Material
Added Prefab of the whole rig, this includes hands and digit colliders.
One collider on index finger that will be active when the point
animation is displayed, will turn off when not pointing.
Added Watch on left hand Only active in dimentions and only displays
time left when touchpad is pressed.
Added Harpic feedbackscript but doesnt work at the moment.